+++ backend/pkg/kafka/config_sasl.go
@@ -6,11 +6,12 @@ import (
 )
 
 const (
-	SASLMechanismPlain       = "PLAIN"
-	SASLMechanismScramSHA256 = "SCRAM-SHA-256"
-	SASLMechanismScramSHA512 = "SCRAM-SHA-512"
-	SASLMechanismGSSAPI      = "GSSAPI"
-	SASLMechanismOAuthBearer = "OAUTHBEARER"
+	SASLMechanismPlain                  = "PLAIN"
+	SASLMechanismScramSHA256            = "SCRAM-SHA-256"
+	SASLMechanismScramSHA512            = "SCRAM-SHA-512"
+	SASLMechanismGSSAPI                 = "GSSAPI"
+	SASLMechanismOAuthBearer            = "OAUTHBEARER"
+	SASLMechanismAWSManagedStreamingIAM = "AWS-MSK-IAM"
 )
 
 // SASLConfig for Kafka client
@@ -19,13 +20,17 @@ type SASLConfig struct {
 	Username     string           `yaml:"username"`
 	Password     string           `yaml:"password"`
 	Mechanism    string           `yaml:"mechanism"`
+	OAUth        SASLOAuthBearer  `yaml:"oauth"`
 	GSSAPIConfig SASLGSSAPIConfig `yaml:"gssapi"`
+	AWSMskIam    SASLAwsMskIam    `yaml:"awsMskIam"`
 }
 
 // RegisterFlags for all sensitive Kafka SASL configs.
 func (c *SASLConfig) RegisterFlags(f *flag.FlagSet) {
 	f.StringVar(&c.Password, "kafka.sasl.password", "", "SASL password")
+	c.OAUth.RegisterFlags(f)
 	c.GSSAPIConfig.RegisterFlags(f)
+	c.AWSMskIam.RegisterFlags(f)
 }
 
 // SetDefaults for SASL Config
@@ -44,7 +49,15 @@ func (c *SASLConfig) Validate() error {
 			return fmt.Errorf("failed to validate gssapi config: %w", err)
 		}
 	case SASLMechanismOAuthBearer:
-		return fmt.Errorf("sasl mechanism '%v' is valid but not yet supported. Please submit an issue if you need it", c.Mechanism)
+		err := c.OAUth.Validate()
+		if err != nil {
+			return fmt.Errorf("failed to validate OAuth Bearer config: %w", err)
+		}
+	case SASLMechanismAWSManagedStreamingIAM:
+		err := c.AWSMskIam.Validate()
+		if err != nil {
+			return fmt.Errorf("failed to validate aws msk iam config: %w", err)
+		}
 	default:
 		return fmt.Errorf("given sasl mechanism '%v' is invalid", c.Mechanism)
 	}