+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - [FEATURE] Proto files can now be provided via the local file system as well
 - [FEATURE] Schema registry support for Protobuf (including auto deserialization)
 - [FEATURE] Consumer group offsets can now be edited (copy offsets from other group, set to start/end/timestamp) or deleted
+- [ENHANCEMENT] Add two new Kafka SASL Mechanisms: OAuthBearer and AWS MSK IAM
 - [ENHANCEMENT] Preview tags in the message viewer can now be named
 - [ENHANCEMENT] Maximum string length in JSON viewer has been increased from 40 to 200 characters
 - [ENHANCEMENT] CTRL + Click on values in JSON viewer will copy the value into the clipboard